Add main.js content and fix html
change html id for begin investigation to make connection between html and main.js because html used a different id on begin investigation button and caused it to have no efffect when the button was clicked. now the button click works but the screen goes blank and nothing else happens. Need to fix.